+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ## [0.0.27] - 2020-06-29
 ### 🐛 Features
-- Add purchase endpoint ([#217](https://github.com/whirli/whirli-client/pull/220))
+- Add purchase endpoint ([#220](https://github.com/whirli/whirli-client/pull/220))
 
 ## [0.0.26] - 2020-06-26
 ### 🐛 Bug fixes