+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### 🐛 Features
+### 🚀 Features
 - Add purchase endpoint ([#220](https://github.com/whirli/whirli-client/pull/220))
 
 ## [0.0.26] - 2020-06-26