+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.27] - 2020-07-02
 ### 🚀 Features
 - Add picking endpoint to orders and gifts resource ([#225](https://github.com/whirli/whirli-client/pull/225))